requirements.txt
RPi.GPIO>=0.7.1
mfrc522>=0.0.7